backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
opencv-python==4.11.0.86
numpy==2.1.3
scikit-image==0.24.0
pandas==2.2.3
openpyxl==3.1.5
pillow==11.0.0
pydantic-settings==2.6.1
